Add Larastan Library
Larastan is a wrapper around the PHPStan static analysis tool which
is tuned specifically for Laravel applications. It is used to find
bugs and violations of best-practices.

We turn the analysis level all the way up to ensure we are always
writing clean code and that bugs are caught early.