Add Foundational Tests
We set up a simple web route and console command and corresponding
tests to ensure that the application is functioning correctly.

These routes and tests will likely be removed later once we have some
actual functionality to test, but for now they serve as a basic
assurance that the core framework is working correctly.